Move processor name editor code to the processor info editor module.
