+++ kunquat/tracker/ui/views/processor/infoeditor.py
@@ -14,7 +14,6 @@
 from PySide.QtCore import *
 from PySide.QtGui import *
 
-from .name import Name
 from .processorupdater import ProcessorUpdater
 
 
@@ -31,5 +30,41 @@ class InfoEditor(QWidget, ProcessorUpdater):
         v.setSpacing(4)
         v.addWidget(self._name, 0, Qt.AlignTop)
         self.setLayout(v)
+
+
+class Name(QWidget, ProcessorUpdater):
+
+    def __init__(self):
+        super().__init__()
+        self._edit = QLineEdit()
+
+        h = QHBoxLayout()
+        h.setContentsMargins(0, 0, 0, 0)
+        h.addWidget(QLabel('Name:'))
+        h.addWidget(self._edit)
+        self.setLayout(h)
+
+    def _on_setup(self):
+        self.register_action('signal_controls', self._update_name)
+        self._update_name()
+        QObject.connect(self._edit, SIGNAL('textEdited(QString)'), self._text_edited)
+
+    def _update_name(self):
+        old_block = self._edit.blockSignals(True)
+        module = self._ui_model.get_module()
+        au = module.get_audio_unit(self._au_id)
+        proc = au.get_processor(self._proc_id)
+        vis_name = proc.get_name() or ''
+        if vis_name != str(self._edit.text()):
+            self._edit.setText(vis_name)
+        self._edit.blockSignals(old_block)
+
+    def _text_edited(self, text):
+        text = str(text)
+        module = self._ui_model.get_module()
+        au = module.get_audio_unit(self._au_id)
+        proc = au.get_processor(self._proc_id)
+        proc.set_name(text)
+        self._updater.signal_update('signal_controls')
